refactor: 2단계 - 로또, 패키지 변경
